requirements.txt
Appium-Python-Client==3.2.1
attrs==23.2.0
behave==1.2.6
browserstack-local==1.2.7
browserstack-sdk==1.19.14
certifi==2024.2.2
charset-normalizer==3.3.2
dukpy==0.3.1
exceptiongroup==1.2.0
execnet==2.0.2
filelock==3.13.1
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
idna==3.6
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
parse==1.20.1
parse-type==0.6.2
percy-appium-app==2.0.4
pluggy==1.4.0
psutil==5.9.8
PyPAC==0.16.4
PySocks==1.7.1
pytest==8.0.2
pytest-xdist==3.5.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
selenium==4.18.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
tldextract==5.1.1
tomli==2.0.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
